Move WESLGrammar to /pars
