+++ tools/packages/wesl/src/parse/WESLGrammar.ts
@@ -19,8 +19,8 @@ import {
   withSep,
   withSepPlus,
 } from "mini-parse";
-import { weslImports } from "./parse/ImportGrammar.ts";
-import { templateClose, templateOpen, WeslToken } from "./parse/WeslStream.ts";
+import { weslImports } from "./ImportGrammar.ts";
+import { templateClose, templateOpen, WeslToken } from "./WeslStream.ts";
 import {
   aliasCollect,
   collectAttribute,
@@ -40,8 +40,8 @@ import {
   stuffCollect,
   typedDecl,
   typeRefCollect,
-} from "./WESLCollect.ts";
-import { mainTokens } from "./WESLTokens.ts";
+} from "../WESLCollect.ts";
+import { mainTokens } from "../WESLTokens.ts";
 
 export const word = kind(mainTokens.ident);
